+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## main
 
+    - Moved install scripts into scripts folder. 
+      - There is a user install script (conda-lock) and a development install script (same as before).
     - Add conda-lock support for pinning environmental dependencies
     - Fix broken links in docs (mostly to github).
     - Pin scipy version to >=1.16.0 and <1.17.0 to resolve Amazon Machine Image build issue.